### 2024-05-25 #6780 @editorialbot
[PRE REVIEW]: dust_extinction: Interstellar Dust Extinction Models
**Submitting author:** @karllark (<a href="http://orcid.org/0000-0001-5340-6774">Karl Gordon</a>)
**Repository:** https://github.com/karllark/dust_extinction
**Branch with paper.md** (empty if default branch):
**Version:** v1.4.1
