### 2024-06-23 #6780 @arfon
☝ to generate the most similar paper recommendation list which has been broken for some time.
