### 2024-07-09 #6780 @karllark
I don't have any particular suggestions.  I can think about it.  Should the suggestions come from those already signed up to be JOSS reviewers or can the be anyone?
