### 2024-07-19 f0c41d1
DEP: refactored out an undeclared runtime dependency on setuptools
