### 2024-07-29 cb29df6
updates based on JOSS review and fixing broken link
