+++ paper/paper.bib
@@ -17,6 +17,18 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@software{Barbary16,
+       author = {{Barbary}, Kyle},
+        title = "{extinction v0.3.0}",
+         year = 2016,
+        month = dec,
+          eid = {10.5281/zenodo.804967},
+          doi = {10.5281/zenodo.804967},
+    publisher = {Zenodo},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2016zndo....804967B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @ARTICLE{Bastiaansen92,
        author = {{Bastiaansen}, P.~A.},
         title = "{Narrow band multicolor photometry of reddened and unreddened early-type stars.}",
