+++ paper/paper.md
@@ -36,19 +36,21 @@ in one place with a consistent interface.
 Many observation- and theory-based extinction curves have been presented
 in the literature. Having one python package providing these models ensures
 that they are straightforward to use and used within their valid wavelength
-and parameter (where appropriate) ranges.
+and parameter (where appropriate) ranges.  Other packages provide extinction
+curves, but they generally provide one or a small number of curves for 
+specialized purposes [e.g., @Barbary16].
 
 The types of extinction models supported are Averages, Parameter Averages, Grain
 Models, and Shapes. The Averages are averages of a set of measured extinction
-curves and examples are shown in Fig. \autoref{fig:averages}. The Parameter
+curves and examples are shown in \autoref{fig:averages}. The Parameter
 Averages are extinction curve averages that depend on a parameter, often $R(V) =
-A(V)/E(B-V)$ which is the ratio of total to selective extinction. Fig.
+A(V)/E(B-V)$ which is the ratio of total to selective extinction. 
 \autoref{fig:parameter_averages} shows examples of such models. The Grain
-Models are those extinction curves computed using dust grain models. See Fig.
+Models are those extinction curves computed using dust grain models. See
 \autoref{fig:grain} and note that these models provide dust extinction
 predictions from X-ray through submm wavelengths. The final type of models
 are Shapes that provide flexible functional forms that fit selected wavelength
-ranges (see Fig. \autoref{fig:shapes} for an example).
+ranges (see \autoref{fig:shapes} for an example).
 
 ![Examples of Average models based on observations in the Milky Way, Large Magellanic Cloud (LMC), and Small Magellanic Cloud (SMC) [@Bastiaansen92; @Gordon03; @Gordon09; @Gordon21; @Gordon24].\label{fig:averages}](average_models_uv_nir.png){
 width=70% }
@@ -68,7 +70,7 @@ a helper `extinguish` function that alternatively provides the fractional
 effects of extinction for a specific dust column (e.g., $A(V)$ value). This
 allows for the effects of dust to be modeled for or removed from an observation.
 
-This package does not implement dust attenuation models. Dust attenuation
+This package does not implement dust attenuation models[^1]. Dust attenuation
 results when observing more complex systems like a star with nearby,
 circumstellar dust or a galaxy with many stars extinguished by different amounts
 of dust. In both cases, the wavelength dependence of effects of dust are
@@ -76,6 +78,8 @@ dependent not just on the dust grain properties, but also the effects of the
 dust radiative transfer [@Steinacker13]. Specifically, these effects are the averaging of sources
 extinguished by differing amount of dust and the inclusion of a significant
 number of photons scattered into the observing beam.
+
+[^1]: See [karllark/dust_attenuation](https://github.com/karllark/dust_attenuation).
 
 Any published dust extinction model is welcome for inclusion in this package.
 
