+++ docs/index.rst
@@ -164,11 +164,9 @@ contributors who will abide by the `Python Software Foundation Code of Conduct
 `Astropy`_.  The following pages will help you get started with contributing
 fixes, code, or documentation (no git or GitHub experience necessary):
 
-* `How to make a code contribution <https://docs.astropy.org/en/stable/development/workflow/development_workflow.html>`_
+* `How to make a code contribution <https://docs.astropy.org/en/latest/index_dev.html>`_
 
 * `Coding Guidelines <https://docs.astropy.org/en/stable/development/codeguide.html>`_
-
-* `Developer Documentation <https://docs.astropy.org/en/latest/index_dev.html>`_
 
 For the complete list of contributors please see the `dust_extinction
 contributors page on Github
