+++ .github/workflows/tox-tests.yml
@@ -101,7 +101,7 @@ jobs:
       run: |
         tox -e py312-cov
     - name: Upload coverage to codecov
-      uses: codecov/codecov-action@v1
+      uses: codecov/codecov-action@v4
       with:
         token: ${{ secrets.CODECOV }}
         fail_ci_if_error: true
