### 2024-08-09 07991d5
fixing more astropy URLs
