+++ docs/index.rst
@@ -39,6 +39,14 @@ User Documentation
    Fitting extinction curves <dust_extinction/fit_extinction.rst>
    References <dust_extinction/references.rst>
 
+User Documentation
+==================
+
+.. toctree::
+   :maxdepth: 2
+
+   Model Base Classes (how to add a model) <dust_extinction/dev_model.rst>
+
 Installation
 ============
 
