+++ docs/dust_extinction/fit_extinction.rst
@@ -1,3 +1,4 @@
+.. _fit_curves:
 #####################
 Fit Extinction Curves
 #####################
