+++ docs/dust_extinction/fit_extinction.rst
@@ -1,4 +1,5 @@
 .. _fit_curves:
+
 #####################
 Fit Extinction Curves
 #####################
