### 2024-08-14 #7023 @karllark
I have updated the paper and the repository based on @TheSkyentist suggestions.
