### 2024-08-14 6ae29a6
updating G24 in JOSS paper
