+++ paper/paper.bib
@@ -256,18 +256,21 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Gordon24,
-       author = {{Gordon}, Karl D. and {Fitzpatrick}, E.~L. and {Massa}, Derck and {Bohlin}, Ralph and {Chastenet}, Jeremy and {Murray}, Claire E. and {Clayton}, Geoffrey C. and {Lennon}, Daniel J. and {Misselt}, Karl A. and {Sandstrom}, Karin},
-        title = "{Expanded Sample of Small Magellanic Cloud Ultraviolet Dust Extinction Curves: Correlations between the 2175 A bump, q\_pah, UV extinction shape, and N(HI)/A(V)}",
-      journal = {arXiv e-prints},
-     keywords = {Astrophysics - Astrophysics of Galaxies},
+       author = {{Gordon}, Karl D. and {Fitzpatrick}, E.~L. and {Massa}, Derck and {Bohlin}, Ralph and {Chastenet}, J{\'e}r{\'e}my and {Murray}, Claire E. and {Clayton}, Geoffrey C. and {Lennon}, Daniel J. and {Misselt}, Karl A. and {Sandstrom}, Karin},
+        title = "{Expanded Sample of Small Magellanic Cloud Ultraviolet Dust Extinction Curves: Correlations between the 2175 {\r{A}} Bump, q $_{PAH}$, Ultraviolet Extinction Shape, and N(H I)/A(V)}",
+      journal = {\apj},
+     keywords = {Interstellar dust, Interstellar dust extinction, Ultraviolet extinction, Small Magellanic Cloud, 836, 837, 1738, 1468, Astrophysics - Astrophysics of Galaxies},
          year = 2024,
-        month = may,
-          eid = {arXiv:2405.12792},
-        pages = {arXiv:2405.12792},
+        month = jul,
+       volume = {970},
+       number = {1},
+          eid = {51},
+        pages = {51},
+          doi = {10.3847/1538-4357/ad4be1},
 archivePrefix = {arXiv},
        eprint = {2405.12792},
  primaryClass = {astro-ph.GA},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2024arXiv240512792G},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024ApJ...970...51G},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
