### 2024-08-15 #7023 @TheSkyentist
The purpose, documentation, and examples for the code are all well described in the JOSS submission. The code is installable, flexible, and useful for current and future astronomical analysis. With the additional changes made by @karllark I would recommend this submission be accepted.
