+++ docs/dust_extinction/dev_model.rst
@@ -13,29 +13,29 @@ implemented models.
 All
 ===
 
+All dust extinction models have at least the following:
+
+* A member variable `x_range` that that define the valid range of wavelengths. These are defined in inverse microns as is common for extinction curve research.
+* A member function `evaluate` that computes the extinction at a given `x` and any model parameter values.  The `x` values are checked to be within the valid `x_range`. The `x` values should have astropy.units.  If they do not, then they are assumed to be in inverse microns and a warning is issued stating such.
+
 All of these classes used in ``dust_extinction`` are based on the 
 `Model <https://docs.astropy.org/en/stable/modeling/>`_ astropy.modeling class.
 See the astropy docs for the full details of all that is possible with this class.
 
-All dust extinction models have the following:
-
-* A member variable `x_range` that that define the valid range of wavelengths. These are defined in inverse microns for historical reasons.
-* A member function `evaluate` that computes the extinction at a given `x` and any model parameter values.  The `x` values are checked to be within the valid `x_range`. The `x` values should have astropy.units.  If they do not, then they are assumed to be in inverse microns and a warning is issued stating such.
-
 FittableModel
 =============
 
-The ``dust_extinction`` shape models are based on the astropy `FittableModel`. 
+The ``dust_extinction`` `shape` models are based on the astropy `FittableModel`. 
 One general use case for these models is to fit observed dust extinction curves.
 See :ref:`fit_curves`.  These models follow the standard astropy setup for such
-models.  This includes defining the parameters to be fit with the `Parameter`
+models.  This includes defining the parameters to be fit with the astropy `Parameter`
 function.
 
 Thus all `shape` models have:
 
-* The member variable `x_range` and function `evaluate` (see :ref:`allmods`)
-* Member parameters that are defined with the astropy `Parameter` function.  This includes default starting values and any standard bounds on the parameters. The number and name of the paramaeters varies by `shape` model.
-* The `evaluate` function that calculates the extinction curve based on the input parameters.
+* The member variable `x_range` and function `evaluate` are set in each `shape` model explicitly.
+* Member parameters are explicitly defined with the astropy `Parameter` function.  This includes default starting values and any standard bounds on the parameters. The number and name of the paramaeters varies by `shape` model.
+* The `evaluate` function that calculates the extinction curve based on the input parameters.  This is explicitly defined in each `shape` model.
 
 BaseExtModel
 ============
@@ -47,9 +47,9 @@ the rest of the `dust_extinction` models.   This model provides the
 All of the `average` models are based on `BaseExtModel` directly.  Thus 
 all the `average` models have:
 
-* The member variable `x_range` and function `evaluate` (see :ref:`allmods`). The `evaluate` function may interpolate the observed average extinction curve or it may be based on a `shape` fit to the observed data.
-* The member function `extinguish`.
-* A member parameter `Rv` that gives the ratio of absolute to selective extinction (i.e., R(V) = A(V)/E(B-V)).  This is not set with the astropy `Parameter` function as is included for reference.
+* The member variable `x_range` and function `evaluate` (see :ref:`allmods`). These are set explicitly for each `average` model.  The `evaluate` function may interpolate the observed average extinction curve or it may be based on a `shape` fit to the observed data.
+* The member function `extinguish` inherited from the `BaseExtModel`.
[…]
