### 2024-08-15 b8d8523
more dev doc info
