+++ docs/dust_extinction/choose_model.rst
@@ -179,6 +179,8 @@ observationally (e.g., in the extreme UV below 912 A).
 +--------------+----------------+------------------+--------------+
 |  HD23 MWRV31 |  0.000033 - 10 |      0.1 - 30000 |  MW R(V)=3.1 |
 +--------------+----------------+------------------+--------------+
+|   Y24 MWRV31 |   0.00001 - 25 |    0.04 - 100000 |  MW R(V)=3.1 |
++--------------+----------------+------------------+--------------+
 
 Shape Models
 ============
