+++ docs/dust_extinction/references.rst
@@ -134,6 +134,9 @@ VCG04: `Valencic, Clayton, & Gordon 2014, 616, 912
 WD01: `Weingartner & Draine 2001, ApJ, 548, 296
 <https://ui.adsabs.harvard.edu/abs/2001ApJ...548..296W>`_
 
+Y24: `Ysard et al. 2024, A&A, 684, 34
+<https://ui.adsabs.harvard.edu/abs/2024A%26A...684A..34Y>`_
+
 ZDA04: `Zubko, Dwek, & Arendt 2004, ApJS, 152, 211
 <https://ui.adsabs.harvard.edu/abs/2004ApJS..152..211Z>`_
 
