### 2024-08-15 f73d8ba
removing min when using bounds
