### 2024-08-16 #7023 @ivastar
Excellent! Thank you @keflavich and @TheSkyentist for the speedy review!

I have a few editorial suggestions which I will put in a comment below and then we will move to the close out tasks.
