### 2024-08-16 #7023 @ivastar
Editorial comments on text:

Line 5: "of single star" --> "of a single star"
Line 6: "line-of-sigh to a star" --> "line-of-sigh to the star"
Line 7: "also know" --> "also known"
Line 10 (and throughout the text): Python should always be capitalized (https://devguide.python.org/documentation/style-guide/)
Line 10: "astropy affiliated" --> "Astropy-affiliated", Astropy refers to the project + hyphenation (https://docs.astropy.org/en/latest/development/style-guide.html)
Line 24: "dust grain models. See Figure 3 and note that these models" --> "dust grain models (see Figure 3). Note also that only these Grain models"
Line 27: "for an example" --> for examples
Line 29: that --> which
Lines 19, 23 and then the figure captions: "Models" is capitalized in the text but not in the figure captions. The docs don't have capitalization, so maybe remove it throughout.

Bibliography:
The short-hands for journal names are not properly rendered so can you please replace all \journalname with the full name of the journal, e.g., \aap --> Astronomy and Astrophysics?
