### 2024-08-16 4e83a45
updates based on JOSS editor review
