+++ paper/paper.md
@@ -21,36 +21,36 @@ bibliography: paper.bib
 
 # Summary
 
-Extinction describes the effects of dust on observations of single star due to
-the dust along the line-of-sight to a star removing flux by absorbing photons
-and scattering photons out of the line-of-sight. The wavelength dependence of
-dust extinction (also know as extinction curves) provides fundamental
-information about the size, composition, and shape of interstellar dust grains.
-In general, models giving the wavelength dependence of extinction are used to
-model or correct the effects of dust on observations. This python astropy
-affiliated package [@Astropy22] provides many of the published extinction models
-in one place with a consistent interface.
+Extinction describes the effects of dust on observations of a single star due
+to the dust along the line-of-sight to the star removing flux by absorbing
+photons and scattering photons out of the line-of-sight. The wavelength
+dependence of dust extinction (also known as extinction curves) provides
+fundamental information about the size, composition, and shape of interstellar
+dust grains. In general, models giving the wavelength dependence of extinction
+are used to model or correct the effects of dust on observations. This Python
+Astropy-affiliated package [@Astropy22] provides many of the published
+extinction models in one place with a consistent interface.
 
 # Statement of need
 
-Many observation- and theory-based extinction curves have been presented
-in the literature. Having one python package providing these models ensures
-that they are straightforward to use and used within their valid wavelength
-and parameter (where appropriate) ranges.  Other packages provide extinction
-curves, but they generally provide one or a small number of curves for 
-specialized purposes [e.g., @Barbary16].
+Many observation- and theory-based extinction curves have been presented in the
+literature. Having one Python package providing these models ensures that they
+are straightforward to use and used within their valid wavelength and parameter
+(where appropriate) ranges. Other packages provide extinction curves, but they
+generally provide one or a small number of curves for specialized purposes
+[e.g., @Barbary16].
 
-The types of extinction models supported are Averages, Parameter Averages, Grain
-Models, and Shapes. The Averages are averages of a set of measured extinction
-curves and examples are shown in \autoref{fig:averages}. The Parameter
-Averages are extinction curve averages that depend on a parameter, often $R(V) =
-A(V)/E(B-V)$ which is the ratio of total to selective extinction. 
-\autoref{fig:parameter_averages} shows examples of such models. The Grain
-Models are those extinction curves computed using dust grain models. See
-\autoref{fig:grain} and note that these models provide dust extinction
-predictions from X-ray through submm wavelengths. The final type of models
-are Shapes that provide flexible functional forms that fit selected wavelength
-ranges (see \autoref{fig:shapes} for an example).
+The types of extinction models supported are Averages, Parameter Averages,
+Grain odels, and Shapes. The Averages are averages of a set of measured
+extinction curves and examples are shown in \autoref{fig:averages}. The
+Parameter Averages are extinction curve averages that depend on a parameter,
+often $R(V) = A(V)/E(B-V)$ which is the ratio of total to selective extinction.
+\autoref{fig:parameter_averages} shows examples of such models. The Grain odels
+are those extinction curves computed using dust grain models
+(\autoref{fig:grain)}. Note that only these models provide dust extinction
+predictions from X-ray through submm wavelengths. The final type of models are
+Shapes that provide flexible functional forms that fit selected wavelength
+ranges (see \autoref{fig:shapes} for examples).
 
[…]
