### 2024-08-16 #7023 @karllark
FYI, one of my favorite reflection nebula to study is NGC 7023.  I am quite pleased that this JOSS review is numbered 7023.  Heh.
