+++ paper/paper.bib
@@ -385,6 +385,24 @@ archivePrefix = "arXiv",
   adsnote = {Provided by the Smithsonian/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Ysard24,
+       author = {{Ysard}, N. and {Jones}, A.~P. and {Guillet}, V. and {Demyk}, K. and {Decleir}, M. and {Verstraete}, L. and {Choubani}, I. and {Miville-Desch{\^e}nes}, M. -A. and {Fanciullo}, L.},
+        title = "{THEMIS 2.0: A self-consistent model for dust extinction, emission, and polarisation}",
+      journal = {Astronomy and Astrophysics},
+     keywords = {polarization, dust, extinction, ISM: general, infrared: ISM, submillimeter: ISM, Astrophysics - Astrophysics of Galaxies},
+         year = 2024,
+        month = apr,
+       volume = {684},
+          eid = {A34},
+        pages = {A34},
+          doi = {10.1051/0004-6361/202348391},
+archivePrefix = {arXiv},
+       eprint = {2401.07739},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024A&A...684A..34Y},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @ARTICLE{Zubko04,
    author = {{Zubko}, V. and {Dwek}, E. and {Arendt}, R.~G.},
     title = "{Interstellar Dust Models Consistent with Extinction, Emission, and Abundance Constraints}",
