+++ README.rst
@@ -17,9 +17,15 @@ Please cite the original papers for the models used.  The papers for each model
 are given `here <https://dust-extinction.readthedocs.io/en/latest/dust_extinction/references.html>`_.
 
 In addition, please cite the latest release of the package given below.
+(Soon to be replaced hopefully with JOSS paper.)
 
 .. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.11235336.svg
   :target: https://doi.org/10.5281/zenodo.11235336
+  :alt: current code version DOI
+
+.. image:: https://joss.theoj.org/papers/42619ff7078a218d14893ff2da4f6ca6/status.svg
+  :target: https://joss.theoj.org/papers/42619ff7078a218d14893ff2da4f6ca6
+  :alt: JOSS status
 
 Build checks/status
 -------------------
