+++ README.rst
@@ -17,7 +17,7 @@ Please cite the original papers for the models used.  The papers for each model
 are given `here <https://dust-extinction.readthedocs.io/en/latest/dust_extinction/references.html>`_.
 
 In addition, please cite the latest release of the package given below.
-(Soon to be replaced hopefully with JOSS paper.)
+(Hopefully soon to be replaced with a JOSS paper.)
 
 .. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.11235336.svg
   :target: https://doi.org/10.5281/zenodo.11235336
