+++ CHANGES.rst
@@ -1,3 +1,8 @@
+1.6.dev (unreleased)
+================
+
+- none yet
+
 1.5 (2024-08-16)
 ================
 
