### 2024-08-16 #7023 @karllark
Package released as v1.5.  Uploaded to pypi.

DOI from zenodo.

Gordon, K. (2024). dust_extinction (v1.5). Zenodo. https://doi.org/10.5281/zenodo.13333814
