### 2024-08-17 8925f4f
updating conda badge
