### 2024-08-21 #7023 @ivastar
Hi @karllark, the Zenodo archive license is "Creative commons" for some reason? Can you please change it to match that of the paper, BSD 3-clause "New" or "Revised" license?
