### 2024-08-21 #7023 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
