### 2024-08-21 #7023 @ivastar
🎉  success! Thank you for submitting to JOSS @karllark! Pleasure working with you on this.

The editor in chief has to push the final button for publication, which he will probably do in the next week or so. But our work here is done.
